Added a switch between kg and gr.
Former-commit-id: d1d342728bb6491a07b722dd928886e970160cb7
Former-commit-id: 94944867505f878dd929c47950818410b85bf5f9

+++ DeviceIntegration/Scales/ScaleMeasurementData.cs
@@ -4,9 +4,9 @@
 	{
 		public MeasurementStatus Status { get; }
 
-		public int WeightGr { get; }
+		public long WeightGr { get; }
 
-		public ScaleMeasurementData(MeasurementStatus status, int weightGr)
+		public ScaleMeasurementData(MeasurementStatus status, long weightGr)
 		{
 			Status = status;
 			WeightGr = weightGr;
